WLT-515: benchmark for create member
